requirements.txt
streamlit==1.29.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.36.0
torch==2.1.0
numpy==1.24.3
rank-bm25==0.2.2
plotly==5.18.0
tqdm==4.66.1
python-dotenv==1.0.0
requests==2.31.0